+++ docs/export.el
@@ -1,0 +1,33 @@
+;; Batch export org-mode files to RST for Sphinx.
+;; Usage (repo root): emacs --batch -l docs/export.el
+;; Org under docs/orgmode/ is the source. RST under docs/source/ is generated.
+(require 'package)
+(add-to-list 'package-archives '("melpa" . "https://melpa.org/packages/") t)
+(package-initialize)
+
+(unless (package-installed-p 'ox-rst)
+  (package-refresh-contents)
+  (package-install 'ox-rst))
+
+(require 'ox-rst)
+(require 'ox-publish)
+
+(setq org-export-with-section-numbers nil)
+(setq org-export-with-toc nil)
+(setq org-export-with-author nil)
+(setq org-export-with-timestamps nil)
+(setq org-rst-headline-underline ?-)
+
+(setq org-publish-project-alist
+      '(("capnp-janet-rst"
+         :base-directory "./docs/orgmode/"
+         :base-extension "org"
+         :publishing-directory "./docs/source/"
+         :publishing-function org-rst-publish-to-rst
+         :recursive t
+         :headline-levels 4
+         :with-toc nil
+         :section-numbers nil
+         :with-author nil)))
+
+(org-publish "capnp-janet-rst" t)